Worked out bugs, Things Compile now. Nearly ready for watch testing!

+++ libraries/ADWatch/Magnet.h
@@ -14,7 +14,7 @@ class Magnet{
 	public:	
 		//Methods
 		Magnet(float);
-		uint8_t getCurHeading(void);	
+		float getCurHeading(void);	
 		uint8_t getHeadingIndex(void);
 		void updateHeading(float);
